refactor: move tests to appropriate places
